Add extra tests and move the loadingComponents to mma/shared folder
